requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Faker>=22.0.0
